Added uniform buffers. Moved some uniforms to said buffers.

+++ include/gui.hpp
@@ -23,7 +23,7 @@ namespace gui {
         GLfloat* pointLightPosition;
         GLfloat* pointLightColor;
 
-        GLint* specularShininess;
+        GLfloat* specularShininess;
         GLfloat* specularStrength;
 
         GLfloat* ambientLightIntensity;
